Refactoring directory structure
This commit:
- Moves everything related to monitoring to live under monitoring/
- Removes ambiguous naming conventions (rootController under controller
directory)
- Moves unit tests to live under test/unit
- Removes weird test naming conventions (root_controller_spec instead of
test/unit/controllers/root.js)
- Change imports to make everything (or almost) work properly

Pairing with @carolinaborim
