+++ monitoring/logstash/logstash.conf
