+++ src/fetchers/equipment.js
