+++ src/metrics/transformers/userInfo.js
