+++ src/parsers/equipment.js
