+++ src/schemas/equipment.js
