+++ test/unit/fetchers/canVariables.js
@@ -1,5 +1,5 @@
-import config from '../../src/config';
-import CanVariablesFetcher from '../../src/fetcher/canVariablesFetcher';
+import config from '../../../src/config';
+import CanVariablesFetcher from '../../../src/fetchers/canVariables';
 import helpers from '../helpers';
 import td from 'testdouble';
