+++ test/unit/fixtures/facadeEquipment.json
