+++ test/unit/fixtures/telemetrySearch.json
